feature #2591
git-svn-id: svn+ssh://svn.amanzi.org/var/lib/svn/trunk/awe@5472 0edadee9-8154-4c59-a661-b5f45207fcbd

+++ org.amanzi.awe.afp/src/org/amanzi/awe/afp/exporters/AfpExporter.java
@@ -318,7 +318,7 @@ public class AfpExporter extends Job {
                                     statistic = StatisticManager.getStatistic(model.getDatasetNode());
                                     float kf1=1f;
                                     if (syTrx){
-                                        kf1=ns.getTrxOfSyGroup(gr)/frequencies.length;
+                                        kf1=(float)ns.getTrxOfSyGroup(gr)/frequencies.length;
                                     }
                                     writeInterferenceForTrx(kf1,sectorNode, syTrx,syTrx?gr:trxNode, intWriters[i], sectorIntValues, rf);
 
@@ -373,7 +373,7 @@ public class AfpExporter extends Job {
                     if (arr.length==0){
                         continue;
                     }
-                    kf2=ns.getTrxOfSyGroup(gr)/arr.length;
+                    kf2=(float)ns.getTrxOfSyGroup(gr)/arr.length;
                 }
                 analyseSy=analyseSy||syTrx;
                 // TODO debug (2nd part of old condition do not necessary...)
